moved the attraction carousel into the molecule folder
